notification_system/requirements.txt
twilio==9.10.0
flask==3.1.2
flask-cors==6.0.2
python-dotenv==1.0.0
requests==2.32.5
